feat(PPSC-1136): unify scan sbom — one command for CPE and purl SBOMs
The CLI now has a single ` + "`" + `armis-cli scan sbom <path>` + "`" + ` command. Users no
longer choose between ` + "`" + `scan sbom` + "`" + ` (VEX-only) and ` + "`" + `scan sbom-cpe` + "`" + `
(findings-only) — the backend picks the right scanner based on the
SBOM's contents and the CLI shows findings by default:

* CPE-shaped SBOMs (asset inventories with explicit CPEs) → CpeSbomScanner
  → NVD findings.
* Purl-shaped SBOMs (npm / NuGet / PyPI application manifests) →
  TrivySbomScanner → deps.dev findings.
* Both paths produce a normalised findings table in the terminal; both
  gate on --fail-on; both write a raw JSON dump to
  .armis/<artifact>-sbom.json (override with --sbom-output).

--vex-output opts into VEX generation. VEX comes from the sibling of
whichever scanner ran (CpeVexGenerator or GrypeVexGenerator). Without
--vex-output the CLI skips the VEX slot entirely.

Deleted:
- internal/cmd/scan_sbom_cpe.go + test
- internal/scan/sbomcpe/ (whole package: driver, unit tests, integration
  tests, ~750 lines)

Added:
- internal/scan/normalized_findings.go — BuildScanResult /
  ConvertNormalizedFindings / IsEmptyFinding / CleanDescription lifted
  from the sbomcpe package so scan/sbom (and future scanners) can share
  them without duplicating ~150 lines.

Modified:
- internal/scan/sbom/sbom.go — driver now fetches normalized findings on
  both paths, downloads the raw JSON dump from whichever results_refs
  key is set (sbom_cpe_results or sbom_results), and only downloads VEX
  when --vex-output was requested. Adds ResultKeySBOMCPE constant with a
  contract test.
- internal/cmd/scan_sbom.go — updated NewScanner call with page limit +
  includeNonExploitable; --vex-output opts into VEX; --sbom-output is
  now the raw-findings path.
